Move sidebar-specific code out of Workspace
