+++ crates/workspace/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "workspace"
+version = "0.1.0"
+edition = "2018"
